requirements.txt
django==4.0.5
pillow==9.1.1
django-phonenumber-field==6.2.0
environs==9.5.0
python-telegram-bot==13.12
phonenumbers==8.12.50